Move almost all logic of updateFork() to KifuTree model
